bundle javascript tests with mocha for the browser
