+++ package.json
@@ -5,10 +5,12 @@
     },
     "devDependencies": {
         "html-webpack-plugin": "^3.2.0",
+        "mocha": "^6.1.4",
+        "mocha-loader": "^2.0.1",
+        "text-encoder": "0.0.4",
         "webpack": "^4.11.1",
         "webpack-cli": "^3.1.1",
-        "webpack-dev-server": "^3.1.0",
-        "text-encoder": "0.0.4"
+        "webpack-dev-server": "^3.1.0"
     },
     "dependencies": {
         "chai": "^4.2.0"
